Updates for `/stats` section
Only include scripts on `index` view

Add default `Hit` ordering

Fix long text not wrapping
